Semester-02-Capstone-Project-Customer-ChatBot/requirements.txt
# Core Framework
streamlit>=1.28.0
gradio>=4.0.0

# AI/ML Libraries
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
faiss-cpu>=1.7.4

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
markdown>=3.5.0

# Vector Store & Embeddings
chromadb>=0.4.0
openai>=1.3.0  # Optional for enhanced embeddings

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
tqdm>=4.66.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# HTTP & API
requests>=2.31.0
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
factory-boy>=3.3.0

# Logging & Monitoring
structlog>=23.2.0
rich>=13.7.0

# Utilities
click>=8.1.0
tenacity>=8.2.0  # For retry logic
validators>=0.22.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
